# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Revenue Rescue AI operations API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardMetrics {
  failedRevenue: number;
  recoveredRevenue: number;
  recoveryRate: number;
  revenueAtRisk: number;
  activeCases: number;
  recoveredThisMonth: number;
  additionalRevenue: number;
}

export interface ActivityItem {
  id: string;
  title: string;
  description: string;
  type: string;
  createdAt: string;
  /** @nullable */
  caseId?: string | null;
}

export interface StrategyPerformance {
  strategy: string;
  attempts: number;
  recovered: number;
  successRate: number;
  revenue: number;
}

export interface Payment {
  id: string;
  customerId: string;
  customerName: string;
  amount: number;
  currency: string;
  status: string;
  failureReason: string;
  paymentMethod: string;
  retryCount: number;
  createdAt: string;
}

export interface SimulationInput {
  /** @minimum 100 */
  amount?: number;
  failureReason?: string;
  paymentMethod?: string;
}

export interface Customer {
  id: string;
  name: string;
  email: string;
  segment: string;
  lifetimeValue: number;
  totalPayments: number;
  successfulPayments: number;
  failedPayments: number;
  previousRecoveryRate: number;
  preferredMethod: string;
  preferredTime?: string;
  riskLevel: string;
}

export interface TimelineEvent {
  id: string;
  label: string;
  detail: string;
  timestamp: string;
  completed: boolean;
}

export interface AgentInsight {
  agent: string;
  status: string;
  summary: string;
  confidence: number;
}

export interface RecoveryCase {
  id: string;
  paymentId: string;
  customerId: string;
  customerName: string;
  amount: number;
  currency: string;
  failureReason: string;
  paymentMethod: string;
  retryCount: number;
  recoveryScore: number;
  recoveryProbability: number;
  riskLevel: string;
  strategy: string;
  recommendedTime: string;
  status: string;
  reasoning: string;
  createdAt: string;
  requiresApproval: boolean;
  /** @nullable */
  approvalStatus?: string | null;
  /** @nullable */
  outcome?: string | null;
  timeline: TimelineEvent[];
  insights: AgentInsight[];
}

export type ApprovalInputDecision = typeof ApprovalInputDecision[keyof typeof ApprovalInputDecision];


export const ApprovalInputDecision = {
  approve: 'approve',
  reject: 'reject',
  modify: 'modify',
} as const;

export interface ApprovalInput {
  decision: ApprovalInputDecision;
  /** @nullable */
  note?: string | null;
}

export type AnalyticsComparison = {
  traditional: number;
  ai: number;
  additionalRevenue: number;
};

export interface TrendPoint {
  label: string;
  failed: number;
  recovered: number;
  rate: number;
}

export interface BreakdownPoint {
  label: string;
  value: number;
  amount: number;
}

export interface Analytics {
  trend: TrendPoint[];
  failureReasons: BreakdownPoint[];
  paymentMethods: BreakdownPoint[];
  comparison: AnalyticsComparison;
}

export interface AgentNode {
  id: string;
  name: string;
  role: string;
  status: string;
  confidence: number;
  processed: number;
}

export interface AgentStatus {
  overall: string;
  updatedAt: string;
  agents: AgentNode[];
}