# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Revenue Rescue AI operations API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: string;
  title: string;
  description: string;
  type: string;
  createdAt: string;
  /** @nullable */
  caseId?: string | null;
}