# lib/api-zod/src/generated/types/agentInsight.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Revenue Rescue AI operations API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentInsight {
  agent: string;
  status: string;
  summary: string;
  confidence: number;
}